+++ packages/provider/src/modules/auth/index.ts
@@ -8,6 +8,7 @@ import { BaseModule } from '../base-module';
 import { createJsonRpcRequestPayload } from '../../core/json-rpc';
 import { transformNewAssertionForServer, transformAssertionForServer } from '../../util/webauthn';
 import { createWebAuthCreateCredentialError, createWebAuthnNotSupportError } from '../../core/sdk-exceptions';
+import { SDKEnvironment } from '../../core/sdk-environment';
 
 type LoginWithMagicLinkEvents = {
   'email-sent': () => void;
@@ -22,9 +23,37 @@ export class AuthModule extends BaseModule {
    * of 15 minutes).
    */
   public loginWithMagicLink(configuration: LoginWithMagicLinkConfiguration) {
-    const { email, showUI = true } = configuration;
-    const requestPayload = createJsonRpcRequestPayload(MagicPayloadMethod.LoginWithMagicLink, [{ email, showUI }]);
+    const { email, showUI = true, redirectURI } = configuration;
+    const requestPayload = createJsonRpcRequestPayload(MagicPayloadMethod.LoginWithMagicLink, [
+      { email, showUI, redirectURI },
+    ]);
     return this.request<string | null, LoginWithMagicLinkEvents>(requestPayload);
+  }
+
+  /**
+   * Log a user in with a special one-time-use credential token. This is
+   * currently used during magic link flows with a configured redirect to
+   * hydrate the user session at the end of the flow. If the flow is successful,
+   * this method will return a Decentralized ID token (with a default lifespan
+   * of 15 minutes).
+   *
+   * If no argument is provided, a credential is automatically parsed from
+   * `window.location.search`.
+   */
+  public loginWithCredential(credentialOrQueryString?: string) {
+    let credentialResolved = credentialOrQueryString ?? '';
+
+    if (!credentialOrQueryString && SDKEnvironment.target === 'web') {
+      credentialResolved = window.location.search;
+
+      // Remove the query from the redirect callback as a precaution.
+      const urlWithoutQuery = window.location.origin + window.location.pathname;
+      window.history.replaceState(null, '', urlWithoutQuery);
+    }
+
+    const requestPayload = createJsonRpcRequestPayload(MagicPayloadMethod.LoginWithCredential, [credentialResolved]);
+
+    return this.request<string | null>(requestPayload);
   }
 
   public async registerWithWebAuthn(configuration: RegisterWithWebAuthnConfiguration) {